fix gymnasium warnings

+++ simulation-system/libs/gym-csle-stopping-game/src/gym_csle_stopping_game/dao/stopping_game_config.py
@@ -111,13 +111,17 @@ class StoppingGameConfig(SimulationEnvInputConfig):
         """
         :return: the attacker's observation space
         """
-        return gym.spaces.Box(low=np.array([0, 0, 0]), high=np.array([self.L, 1, 2]), dtype=np.float64, shape=(3, ))
+        return gym.spaces.Box(low=np.array([np.float64(0), np.float64(0), np.float64(0)]),
+                              high=np.array([np.float64(self.L), np.float64(1), np.float64(2)]),
+                              dtype=np.float64, shape=(3,))
 
     def defender_observation_space(self) -> gym.spaces.Box:
         """
         :return: the defender's observation space
         """
-        return gym.spaces.Box(low=np.array([0, 0]), high=np.array([self.L, 1]), dtype=np.float64, shape=(2, ))
+        return gym.spaces.Box(low=np.array([np.float64(0), np.float64(0)]),
+                              high=np.array([np.float64(self.L), np.float64(1)]),
+                              dtype=np.float64, shape=(2,))
 
     def attacker_action_space(self) -> gym.spaces.Discrete:
         """
